consumer-service/requirements.txt

flask
kafka-python
mysql-connector-python
